requirements.txt
flask
pymongo
python-dotenv
google-generativeai
streamlit
json
gunicorn